Renamed "globals.py" to "global_vars.py" to avoid name clash with the builtin "globals()" function.
