+++ kicost/distributors/tme/tme.py
@@ -32,8 +32,8 @@ import json
 from bs4 import BeautifulSoup
 import http.client # For web scraping exceptions.
 from .. import fake_browser
-from ...globals import PartHtmlError
-from ...globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE, DEBUG_HTTP_RESPONSES
+from ...global_vars import PartHtmlError
+from ...global_vars import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE, DEBUG_HTTP_RESPONSES
 
 from .. import distributor
 from ..global_vars import distributor_dict
