+++ kicost/global_vars.py
